Move the structure logging code to logger.go file
